api/requirements.txt
Flask
Werkzeug
requests
python-dotenv